# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /org:
    get:
      operationId: getOrganization
      tags: [organization]
      summary: Get the caller's organization (Clerk session) with its venues
      description: |
        Requires a signed-in Clerk user with an active organization. The
        organization is the billing tenant: it owns the plan, the credit
        balance, and every venue listed here.
      responses:
        "200":
          description: Organization summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: No active organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /org/billing/checkout:
    post:
      operationId: createOrgBillingCheckout
      tags: [organization]
      summary: Start Stripe Checkout for the caller's organization
      description: |
        Subscriptions (starter/growth) and one-off credit packs are purchased
        by the organization; credits land in the shared org balance via the
        Stripe webhook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingCheckoutResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: No active organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Billing not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /org/billing/portal:
    post:
      operationId: createOrgBillingPortal
      tags: [organization]
      summary: Open the Stripe customer portal for the caller's organization
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPortalResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: No active organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Billing not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /org/credit-history:
    get:
      operationId: getOrgCreditHistory
      tags: [organization]
      summary: Recent credit ledger entries for the caller's organization
      responses:
        "200":
          description: Ledger entries, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgCreditHistoryResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: No active organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BillingCheckoutBody:
      type: object
      required: [product]
      properties:
        product:
          type: string
          enum: [starter, growth, credit_pack]
    BillingCheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
    BillingPortalResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
...
    OrganizationResponse:
      type: object
      required: [organization, venues]
      properties:
        organization:
          type: object
          required: [id, name, plan, creditsBalance, clerkOrgId]
          properties:
            id:
              type: integer
            name:
              type: string
            plan:
              type: string
              enum: [trial, starter, growth, none]
            creditsBalance:
              type: integer
            billingPeriodEnd:
              type: string
              format: date-time
              nullable: true
            clerkOrgId:
              type: string
            role:
              type: string
              nullable: true
              description: Caller's Clerk role in this organization (e.g. org:admin).
            billingConfigured:
              type: boolean
              description: Whether Stripe billing is configured on this server.
        venues:
          type: array
          items:
            type: object
            required: [id, name, slug, createdAt]
            properties:
              id:
                type: integer
              name:
                type: string
              slug:
                type: string
              tagline:
                type: string
                nullable: true
              createdAt:
                type: string
                format: date-time
    OrgCreditHistoryResponse:
      type: object
      required: [transactions]
      properties:
        transactions:
          type: array
          items:
            type: object
            required: [id, delta, reason, createdAt]
            properties:
              id:
                type: integer
              delta:
                type: integer
              reason:
                type: string
              venueId:
                type: integer
                nullable: true
              sessionId:
                type: integer
                nullable: true
              createdAt:
                type: string
                format: date-time
...